Update Question Catalog Class, question schema and testing JSON
- Update question catalog schema
- Fixing validation errors
- Update validation JSON
- Implementing formatting comments
- Move class QuestionCatalogChange to the QuestionCatalogReader
