+++ assets/advanced_question_catalog/definition.json
@@ -122,7 +122,6 @@
             },
             "osm_element": "OpenWay"
          }
-      ],
-      "isProfessional": true
+      ]
    }
 ]
