+++ lib/api/app_worker/app_worker_interface.dart
@@ -6,7 +6,7 @@ import 'package:flutter_map/flutter_map.dart';
 import 'package:flutter_mvvm_architecture/base.dart';
 import 'package:get_it/get_it.dart';
 
-import '/models/question_catalog/question_catalog.dart';
+import '/models/question_catalog/question_catalog_reader.dart';
 import '/models/element_variants/element_identifier.dart';
 import '/models/authenticated_user.dart';
 import '/models/answer.dart';
